Finished routes implementation

+++ migrations/2020-12-18-170857_add_comments/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE comments (
-    id SERIAL NOT NULL PRIMARY KEY,
-    post_id INT NOT NULL,
+    id BIGSERIAL NOT NULL PRIMARY KEY,
+    post_id BIGINT NOT NULL,
     content TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL,
     CONSTRAINT fk_post
